Represent ImageOnlyFailures with a separate icon when looking at results on the databases https://bugs.webkit.org/show_bug.cgi?id=248911 rdar://102920285
Reviewed by NOBODY (OOPS!).

It would be easier if ImageOnlyFailures were marked with a different
icon from text failures. The changes in this commit are to effectivly
mark ImageOnlyFailures with a blue I instead of a red X. And the colour
blue has been specifically chosen as one that is colour blind friendly.

* Tools/Scripts/libraries/resultsdbpy/resultsdbpy/view/static/js/expectations.js:
(Expectations.colorMap):
(Expectations.unexpectedResults):
(Expectations):
* Tools/Scripts/libraries/resultsdbpy/resultsdbpy/view/static/library/css/webkit.css:
(:root):