[JSC] Make Move32 elimination more optimistic https://bugs.webkit.org/show_bug.cgi?id=249086 rdar://103218582
Reviewed by Justin Michaud.

We are not removing Move32 if arg0's def is more than 32bit or arg1's use is more than 32bit.
But we do not need to make it super conservative. This patch relaxes this condition so that
we avoid optimization only when the arg0's def is more than 32bit. So, we can allow removing
Move32 where source-def is 32bit and destination-use is 64bit.
This improves JetStream2 by 0.45% with 98% confidence.

* Source/JavaScriptCore/b3/air/AirAllocateRegistersByGraphColoring.cpp:

Canonical link: https://commits.webkit.org/257694@main