Allow grid shorthands to set masonry to rows and columns. https://bugs.webkit.org/show_bug.cgi?id=249108 rdar://103231154
Reviewed by Matt Woodrow.

CSS grid has the "grid" and "grid-template" properties that can be used
to set the longhands of "grid-template-rows" and
"grid-template-columns." These shorthands should be able to set both
longhand properties to masonry if specified.

This can be done by modifiny consumeGridTrackList to check if the passed
in values are "masonry." Since createGridTrackList already checks for
the scenario where masonry is specified, the rest of the logic should
be able to remain the same.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/parsing/masonry-parsing-expected.txt:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
(WebCore::CSSPropertyParserHelpers::consumeGridTrackList):

Canonical link: https://commits.webkit.org/257713@main