Unreviewed clang/libstdc++ build fix in CSSCustomPropertyValue https://bugs.webkit.org/show_bug.cgi?id=249136
Unreviewed build fix for the Clang and libstdc++ combination.

257701@main added the RefPtr<StyleImage> type to the CSSCustomPropertyValue's
SyntaxValue variant which ended tripping up construction of the VariantValue
variant with either CSSVariableReferenceValue or CSSVariableData Refs. This
seems to be specific to the libstdc++ std::variant<> implementation when used
alongside Clang. Workaround is to avoid the simplest conversion constructor of
the variant and instead use std::in_place_type<T> to explictly state the
intended type that's to be constructed and contained in the variant object.

* Source/WebCore/css/CSSCustomPropertyValue.h:

Canonical link: https://commits.webkit.org/257727@main