Prevent the new process from shutting down in WebPageProxy::triggerBrowsingContextGroupSwitchForNavigation https://bugs.webkit.org/show_bug.cgi?id=249239
Reviewed by Chris Dumez.

In 256488@main I introduced a time between when we decide what process to use for a navigation
and when we call WebPageProxy::continueNavigationInNewProcess.  We need to prevent the process
from shutting down or being put in the process cache during this time or we will hit the release
assertion in WebPageProxy::continueNavigationInNewProcess.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::triggerBrowsingContextGroupSwitchForNavigation):

Canonical link: https://commits.webkit.org/257809@main