AX: Undo unnecessary TextMarker workaround for empty text fields. https://bugs.webkit.org/show_bug.cgi?id=249162 <rdar://problem/103263327>
Reviewed by Chris Fleizach.

The workaround was introduced for https://bugs.webkit.org/show_bug.cgi?id=155613, but it is no longer necessary for VoiceOver and creates problems for the refactoring of AXTextMarker and AXTextMarkerRange.
The range for an empty text field should be a valid, 0-length range, which is what we return now with this change.
Renamed the test to reflect what the test is actually doing.

* LayoutTests/accessibility/mac/text-marker-range-for-empty-contenteditable-expected.txt: Renamed from LayoutTests/accessibility/mac/text-marker-range-for-node-without-children-expected.txt.
* LayoutTests/accessibility/mac/text-marker-range-for-empty-contenteditable.html: Renamed from LayoutTests/accessibility/mac/text-marker-range-for-node-without-children.html.
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::boundaryPoint):
(WebCore::setRangeStartOrEndWithCharacterOffset):
(WebCore::AXObjectCache::rangeForUnorderedCharacterOffsets):

Canonical link: https://commits.webkit.org/257845@main