Protect against RemoteLayerTreeDrawingAreaProxy getting destroyed during waitForAndDispatchImmediately() call https://bugs.webkit.org/show_bug.cgi?id=248660 rdar://101806758
Reviewed by Wenson Hsieh and Ryan Haddad.

In RemoteLayerTreeDrawingAreaProxy::waitForDidUpdateActivityState(), during the execution of
IPC::Connection::waitForAndDispatchImmediately(), `this` might get destroyed as a result of
processing an incoming IPC from the WebProcess. For this reason, we need to check that `this`
is still alive after the call returns.

Note that we cannot easily protect `this` since this is not a RefCounted object.

* Source/WebKit/UIProcess/RemoteLayerTree/RemoteLayerTreeDrawingAreaProxy.mm:
(WebKit::RemoteLayerTreeDrawingAreaProxy::waitForDidUpdateActivityState):

Canonical link: https://commits.webkit.org/252432.945@safari-7614-branch