[JSC] Enable AVX JIT code generation https://bugs.webkit.org/show_bug.cgi?id=249322 rdar://problem/103361339
Reviewed by Justin Michaud.

This patch flips a flag so that x86_64 JIT will use AVX if AVX is supported.
This is a part of bringup work of Wasm SIMD on x86_64 where we support SIMD
operation with AVX.

A/B test says neutral on iMac20,1 and MacBookAir8,2 for JetStream2 and Speedometer2.

* Source/JavaScriptCore/assembler/MacroAssemblerX86Common.h:
(JSC::MacroAssemblerX86Common::supportsAVX):

Canonical link: https://commits.webkit.org/257884@main