REGRESSION(257861@main?): 8X Scrolling API-Tests are constantly failing or timing out https://bugs.webkit.org/show_bug.cgi?id=249339 rdar://103373219
Reviewed by Aditya Keerthi.

Some of these API tests began failing because we're attempting to wait for `KeyboardWillShow`
notifications in the case where the keyboard never actually shows up (due to the fact that API tests
run in a non-UI-app context). Mitigate this by checking for the existence of an `UIKeyboard` right
after showing the keyboard, after focusing an editable element.

* Source/WebKit/Platform/spi/ios/UIKitSPI.h:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _didFinishTextInteractionInTextInputContext:]):

Additionally, some of the `RequestTextInputContext` API tests are legitimately failing due to the
fact that they expect us to zoom to reveal the focused element after finishing interaction in a
`_WKTextInputContext`; the way this worked before is that we set the
`m_waitingForPostLayoutEditorStateUpdateAfterFocusingElement` flag, which would cause us to then
call into `-_zoomToRevealFocusedElement` after the next editor state update.

To fix this, we create a new `_revealFocusedElementDeferrer` in
`-_didFinishTextInteractionInTextInputContext:`, allowing us to reveal the focused element after the
next editor state arrives.

(-[WKContentView _elementDidFocus:userIsInteracting:blurPreviousNode:activityStateChanges:userObject:]):

Canonical link: https://commits.webkit.org/257893@main